Added a lot of the scene generation code

+++ data_classes/Light.py
@@ -3,6 +3,6 @@ from data_classes.Point import Point
 
 class Light:
 
-    def __init__(self, position: Point, intensity: int):
+    def __init__(self, position: Point, intensity: float):
         self.position = position
         self.intensity = intensity
