+++ data_classes/Point.py
@@ -1,6 +1,6 @@
 class Point:
 
-    def __init__(self, x: int, y: int, z: int):
+    def __init__(self, x: float, y: float, z: float):
         self.x = x
         self.y = y
         self.z = z
